Moved common methods to utility class.
